ServoPID: Implement C# part of arduino protocol + cleanup C++ code

+++ ServoPID/TestServoPID/ArduinoMock.cpp
@@ -2,6 +2,6 @@
 
 // ReSharper disable CppInconsistentNaming
 
-int gMicros = 0;
+unsigned long gMicros = 0;
 std::vector<int> gAnalogPins(256, 0);
 MockSerial Serial;
